Add files via upload
Using this node, we can route the message to output terminal if and only if a particular field or field data in present in the message. This we can set using XPATH expression.
We can add multiple output terminals to this node. Failure and Default are static terminals to this node.
if all filter patterns return false, then the message is routed to Default Terminal.
if while filtering any failure occurs then message is router to Failure Terminal.
You can set Distribution mode to All, if you want to route the message to all filter patterns which return true.
You can set Distribution mode to First, if you want to route the message to first filter pattern which return true even though others return true in the order.